+++ src/lib.rs
@@ -28,6 +28,7 @@ mod compression;
 pub mod config;
 mod config_access;
 mod config_admin;
+mod config_cache;
 mod config_compression;
 mod config_header;
 mod config_http;